Add a decoupled SessionFactory
