+++ src/java.base/share/classes/java/lang/foreign/Arena.java
@@ -26,6 +26,7 @@
 package java.lang.foreign;
 
 import jdk.internal.foreign.MemorySessionImpl;
+import jdk.internal.foreign.SessionFactory;
 import jdk.internal.javac.PreviewFeature;
 
 import java.util.Objects;
@@ -135,7 +136,7 @@ public interface Arena extends SegmentAllocator, AutoCloseable {
      * @return a new confined arena.
      */
     static Arena openConfined() {
-        return makeArena(MemorySessionImpl.createConfined(Thread.currentThread()));
+        return makeArena(SessionFactory.createConfined(Thread.currentThread()));
     }
 
     /**
@@ -143,7 +144,7 @@ public interface Arena extends SegmentAllocator, AutoCloseable {
      * @return a new shared arena.
      */
     static Arena openShared() {
-        return makeArena(MemorySessionImpl.createShared());
+        return makeArena(SessionFactory.createShared());
     }
 
     private static Arena makeArena(MemorySessionImpl sessionImpl) {
